requirements.txt
pytorch-fid
tqdm
opencv-python
omegaconf
plyfile
lpips
pyquaternion
matplotlib==3.5.3
Cython
scikit-learn
numpy==1.22.4
nuscenes-devkit
pykitti
tensorboardX
tensorboard
diffusers==0.27.2
albumentations==0.5.2
webdataset
transformers==4.39.3
huggingface_hub==0.20.2
easydict
pytorch_lightning==1.5.0
kornia==0.5.0
timm==0.4.12
einops
open_clip_torch==2.16.0
yacs==0.1.8
mediapy
open3d
trimesh